Update build process (#26)
* Specifying DEFAULT_LIBCLANG_PATH

* Setting up default LIBCLANG location

* Setting default LIBCLANG to 10

* Adding win32 libs

* Adding lib default values for windows

* Remove clang configuration from build.rs

* Readd missing changes

* Add android libraries

* Condition lib copy to arch

* Refacto macos dylib management

* Fixing target libraries copy

* Insuring arch and lib matches.

* Changing blind cast to TryFrom, because some arch are not matching usize <=> u64, also focre to remove const attribute.

* Simplify build.rs

* Stripping android libraries

* Revert linux lib files

* Fix linux lib paths

* Revert linux lib update

* downgrade bindgen

* Add missing sym file

* wrong target_os

* Refactoring lib orga and fixing build

* Updating libcblite from 3.0.1 to 3.0.2

* Fixing some version items

* Setting up ubuntu image to 20.04 focal LTS

* Changing LIBCLANG_PATH according to LLVM 10

* Refactoring and fixing build.rs for cross compilation.

* Checking if arm is instead armv7

* Stripping android binaries to lighten release

* Removing libcblite-3.0.1

* Adding ios framework

* Use Result return for all functions in build.rs

* Remove test of fix CB bug

* Simplifying libs files

* Fixing lib copy

* Trial to config ios

* Pointing ios specific headers

* Fix ios build

* Add wrapper_ios

* Cancelling custom inclusion path for ios

* Fix replicator problem endpoint url without slash

* Use include headers for ios

* Copy ios framework to target

* Copy xcframework directory

* Adding libicu v66 & fixing bindings.

+++ src/blob.rs
@@ -174,7 +174,7 @@ impl<'r> std::io::Read for BlobReader<'r> {
                 CBLBlobReader_Read(
                     self.get_ref(),
                     buf.as_mut_ptr().cast::<c_void>(),
-                    buf.len() as u64,
+                    buf.len(),
                     err,
                 )
             })
@@ -229,7 +229,7 @@ impl<'r> std::io::Write for BlobWriter<'r> {
                 let ok = CBLBlobWriter_Write(
                     self.get_ref(),
                     data.as_ptr().cast::<c_void>(),
-                    data.len() as u64,
+                    data.len(),
                     err,
                 );
                 if ok {
